app/requirements.txt
PyChromecast==14.0.5
zeroconf==0.146.1
requests==2.32.3
ffmpeg==1.4
pydub==0.25.1
ffmpeg-python==0.2.0
DateTime==5.5
python-dateutil==2.9.0.post0